introduces random_from_set rule
