+++ tests/test_rules/test_relational_rules/test_row_expression_rules.py
@@ -52,7 +52,8 @@ def test_derived_from_multiple_columns(employees_df, test_seed):
     for row in output.collect():
         assert row["first_name"] + " " + row["last_name"] == row["full_name"]
 
-    assert "first_name" in rule.collect_parent_columns() 
+    assert "first_name" in rule.collect_parent_columns()
+
 
 def test_compatible_with_case_when_expression(sales_df, test_seed):
     rule = RowExpressionRule(
@@ -70,11 +71,11 @@ def test_compatible_with_case_when_expression(sales_df, test_seed):
     runtime = MockRuntime()
 
     output = rule.apply(
-        runtime=cast(Any, runtime), base_df=sales_df, target_col="customer_feedback_score"
+        runtime=cast(Any, runtime),
+        base_df=sales_df,
+        target_col="customer_feedback_score",
     )
 
     for row in output.collect():
         assert row["customer_feedback_score"] in [2.0, 3.0]
-        assert "delivery_distance_miles" in rule.collect_parent_columns() 
-
-
+        assert "delivery_distance_miles" in rule.collect_parent_columns()
